♻️ refactor(theme): split theme logic into feature layers
