+++ src/features/theme/index.ts
@@ -1,0 +1,3 @@
+export type { ResolvedTheme, ThemeMode } from "./lib/theme";
+export { useTheme } from "./model/use-theme";
+export { ThemeSync } from "./ui/theme-sync";
