+++ src/lib/constants.ts
@@ -1,0 +1,3 @@
+export const storageKey = {
+  theme: "theme",
+};
